+++ src/moves/StandardAttack.sol
@@ -51,7 +51,7 @@ contract StandardAttack is IMoveSet, Ownable {
         uint256 attackerPlayerIndex,
         uint256,
         uint256 defenderMonIndex,
-        uint240,
+        uint16,
         uint256 rng
     ) public virtual {
         _move(engine, battleKey, attackerPlayerIndex, defenderMonIndex, rng);
@@ -83,7 +83,7 @@ contract StandardAttack is IMoveSet, Ownable {
         );
     }
 
-    function isValidTarget(IEngine, bytes32, uint240) public pure returns (bool) {
+    function isValidTarget(IEngine, bytes32, uint16) public pure returns (bool) {
         return true;
     }
