+++ test/mocks/GlobalEffectAttack.sol
@@ -33,7 +33,7 @@ contract GlobalEffectAttack is IMoveSet {
         return "Effect Attack";
     }
 
-    function move(IEngine engine, bytes32, uint256 attackerPlayerIndex, uint256, uint256, uint240, uint256) external {
+    function move(IEngine engine, bytes32, uint256 attackerPlayerIndex, uint256, uint256, uint16, uint256) external {
         engine.addEffect(2, 0, EFFECT, bytes32(attackerPlayerIndex));
     }
 
@@ -49,7 +49,7 @@ contract GlobalEffectAttack is IMoveSet {
         return TYPE;
     }
 
-    function isValidTarget(IEngine, bytes32, uint240) external pure returns (bool) {
+    function isValidTarget(IEngine, bytes32, uint16) external pure returns (bool) {
         return true;
     }
